tools/sql-tests/requirements.txt
psycopg2-binary
jinja2